backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
dnspython==2.6.1
litellm==1.48.0
websockets==13.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
